add results

+++ C_module/src/time_evol.c
@@ -7,7 +7,7 @@
 #include "matrix_operations.h"
 #include "time_evol.h"
 void Time_evol(double complex *state, double tf, double dt, double temp, double j_0_3, double j_0_4){
-    int N = (int) (tf/dt), i;
+    long N = (long) (tf/dt), i;
     for(i = 0; i < N; i++){
         RK2_step(state, dt, temp, j_0_3, j_0_4);
     }
